Добавлены PK BIGSERIAL для join-таблиц для облегчения тестирования;                         Изменены методы JoinDaoImpl для возвращения созданных id;                         Написан тест JoinDaoImpl

+++ bookshelf/src/main/java/ru/otus/homework/dao/JoinDao.java
@@ -2,8 +2,8 @@ package ru.otus.homework.dao;
 
 public interface JoinDao {
 
-    int bookAuthorBindingCreate(long bookId, long authorId);
-    int bookGenreBindingCreate(long bookId, long genreId);
+    long bookAuthorBindingCreate(long bookId, long authorId);
+    long bookGenreBindingCreate(long bookId, long genreId);
     int bookBindingsDelete(long bookId);
 
 }
